Moved each function around in the module tree
