+++ src/board/mod.rs
@@ -1,0 +1,7 @@
+mod chessboard;
+mod displays;
+mod movement;
+mod piece;
+pub mod position;
+
+pub use chessboard::Chessboard;
